[AOT Refactor] schemas
[ghstack-poisoned]
